+++ desktop/PomeloApp/Package.swift
@@ -14,6 +14,9 @@ let package = Package(
         .package(path: "LocalPackages/CodeEditSourceEditor"),
         .package(path: "LocalPackages/CodeEditLanguages"),
         .package(path: "LocalPackages/CodeEditTextView"),
+        .package(path: "../../shared/PomeloTerminalKit"),
+        .package(path: "../../shared/PomeloUI"),
+        .package(path: "../../shared/PomeloCore"),
     ],
     targets: [
         .target(name: "CPom"),
@@ -23,7 +26,10 @@ let package = Package(
                            .product(name: "Grape", package: "Grape"),
                            .product(name: "CodeEditSourceEditor", package: "CodeEditSourceEditor"),
                            .product(name: "CodeEditLanguages", package: "CodeEditLanguages"),
-                           .product(name: "CodeEditTextView", package: "CodeEditTextView")],
+                           .product(name: "CodeEditTextView", package: "CodeEditTextView"),
+                           .product(name: "PomeloTerminalKit", package: "PomeloTerminalKit"),
+                           .product(name: "PomeloUI", package: "PomeloUI"),
+                           .product(name: "PomeloCore", package: "PomeloCore")],
             resources: [.process("Resources/Assets.xcassets")],
             linkerSettings: [
                 .unsafeFlags([